octofit-tracker/backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==4.5.0
djongo==1.3.6
pymongo==3.12
sqlparse==0.2.4
